requirements.txt
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flower==1.0.0
pyquery==1.4.3
python-decouple==3.6
python-slugify==6.1.2
redis==4.3.3
requests==2.27.1
selenium==4.2.0
SQLAlchemy==1.4.37
werkzeug==2.1.2
